#2193 integrate swt pie charts in taxonomy viewers
