+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/util/chart/PieChart.java
@@ -5,9 +5,11 @@ import java.util.List;
 import org.eclipse.jface.resource.FontDescriptor;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.PaintEvent;
+import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Font;
 import org.eclipse.swt.graphics.GC;
 import org.eclipse.swt.graphics.Point;
+import org.eclipse.swt.graphics.RGB;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Event;
 import org.eclipse.swt.widgets.Listener;
@@ -21,29 +23,41 @@ import org.eclipse.swtchart.model.Node;
 import name.abuchen.portfolio.money.Values;
 import name.abuchen.portfolio.ui.UIConstants;
 import name.abuchen.portfolio.ui.util.Colors;
+import name.abuchen.portfolio.ui.views.charts.IPieChart;
 
 public class PieChart extends Chart // NOSONAR
 {
     private PieChartToolTip tooltip;
-    
-    public PieChart(Composite parent)
+    private IPieChart.ChartType chartType;
+    private ILabelProvider labelProvider;
+
+    public PieChart(Composite parent, IPieChart.ChartType chartType)
+    {
+        this(parent, chartType, new DefaultLabelProvider());
+    }
+
+    public PieChart(Composite parent, IPieChart.ChartType chartType, ILabelProvider labelProvider)
     {
         super(parent, SWT.NONE);
+        this.chartType = chartType;
+        this.labelProvider = labelProvider;
 
         setData(UIConstants.CSS.CLASS_NAME, "chart"); //$NON-NLS-1$ 
 
-        addListener(SWT.Paint, new Listener()
-        {
-            @Override
-            public void handleEvent(Event event)
+        if (IPieChart.ChartType.DONUT == chartType) {
+            addListener(SWT.Paint, new Listener()
             {
-                // Set color in root node to background color
-                if (getSeriesSet().getSeries().length > 0) {
-                    ICircularSeries<?> cs = (ICircularSeries<?>) getSeriesSet().getSeries()[0];
-                    cs.getRootNode().setColor(getPlotArea().getBackground());
+                @Override
+                public void handleEvent(Event event)
+                {
+                    // Set color in root node to background color
+                    if (getSeriesSet().getSeries().length > 0) {
+                        ICircularSeries<?> cs = (ICircularSeries<?>) getSeriesSet().getSeries()[0];
+                        cs.getRootNode().setColor(getPlotArea().getBackground());
+                    }
                 }
-            }
-        });
+            });
+        }
 
         getPlotArea().addCustomPaintListener(new ICustomPaintListener()
         {
@@ -63,6 +77,18 @@ public class PieChart extends Chart // NOSONAR
         return tooltip;
     }
 
+    /**
+     * Allow to override pie slide label
+     * @param labelProvider
+     */
+    public void setLabelProvider(ILabelProvider labelProvider)
+    {
+        if (labelProvider == null) {
+            return;
+        }
+        this.labelProvider = labelProvider;
+    }
+
     protected void renderLabels(PaintEvent e)
     {
         for(ISeries<?> series : getSeriesSet().getSeries()) {
@@ -73,7 +99,7 @@ public class PieChart extends Chart // NOSONAR
                 List<Node> nodes = ((ICircularSeries<?>)series).getRootNode().getChildren();
                 if (!nodes.isEmpty()) {
                     for(Node node : nodes) {
-                        renderNodeLabel(node, e.gc, xAxis, yAxis);
+                        renderNodeLabel(node, (ICircularSeries<?>) series, e.gc, xAxis, yAxis);
                     }
                 }
             }
@@ -81,9 +107,18 @@ public class PieChart extends Chart // NOSONAR
         
     }
 
-    protected void renderNodeLabel(Node node, GC gc, IAxis xAxis, IAxis yAxis)
+    protected void renderNodeLabel(Node node, ICircularSeries<?> series, GC gc, IAxis xAxis, IAxis yAxis)
     {
-        int level = 2; // fixed for first doughnut level
+        // children drawn first as parent overrides it's section of drawing
+        if(!node.getChildren().isEmpty()) {
+            for(Node childNode : node.getChildren()) {
+                renderNodeLabel(childNode, series, gc, xAxis, yAxis);
+            }
+        }
+        if(node.isVisible() == false)
+            return;
+
+        int level = node.getLevel() - series.getRootNode().getLevel() + (chartType == IPieChart.ChartType.PIE ? 0 : 1);
 
         Font oldFont = gc.getFont();
         gc.setForeground(Colors.WHITE);
@@ -96,13 +131,11 @@ public class PieChart extends Chart // NOSONAR
         FontDescriptor boldDescriptor = FontDescriptor.createFrom(gc.getFont()).setHeight(9);
         gc.setFont(boldDescriptor.createFont(getDisplay()));
 
-        double percent =  ((double)node.getAngleBounds().y / 360);
-        if (percent > 0.025) {
-            String percentString = Values.Percent2.format(percent);
-            Point textSize = gc.textExtent(percentString);
-            gc.drawString(percentString, outerEnd.x - (textSize.x/2) , outerEnd.y - (textSize.y/2), true);
+        String label = labelProvider.getLabel(node);
+        if (label != null) {
+            Point textSize = gc.textExtent(label);
+            gc.drawString(label, outerEnd.x - (textSize.x/2) , outerEnd.y - (textSize.y/2), true);
         }
-
         gc.setFont(oldFont);
     }
 
@@ -113,5 +146,43 @@ public class PieChart extends Chart // NOSONAR
         int xPixelCoordinate = xAxis.getPixelCoordinate(xCoordinate);
         int yPixelCoordinate = yAxis.getPixelCoordinate(yCoordinate);
         return new Point(xPixelCoordinate, yPixelCoordinate);
+    }
+
+    public interface ILabelProvider
+    {
+        public String getLabel(Node node);
+    }
+
+    private static class DefaultLabelProvider implements ILabelProvider
+    {
+        @Override
+        public String getLabel(Node node)
+        {
+            String percentString = null;
+            double percent =  ((double)node.getAngleBounds().y / 360);
+            if (percent > 0.025) {
+                percentString = Values.Percent2.format(percent);
+            }
+            return percentString;
+        }
+    }
+
+    public static final class PieColors
+    {
+        private static final int SIZE = 11;
+        private static final float STEP = 360.0f / (float) SIZE;
+
+        private static final float HUE = 262.3f;
+        private static final float SATURATION = 0.464f;
+        private static final float BRIGHTNESS = 0.886f;
+
+        private int nextSlice = 0;
+
+        public Color next()
+        {
+            float brightness = Math.min(1.0f, BRIGHTNESS + (0.05f * (nextSlice / (float) SIZE)));
+            return Colors.getColor(new RGB((HUE + (STEP * nextSlice++)) % 360f, SATURATION, brightness));
+
+        }
     }
 }
