+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/views/charts/IPieChart.java
@@ -7,6 +7,8 @@ import name.abuchen.portfolio.snapshot.ClientSnapshot;
 
 public interface IPieChart
 {
+    enum ChartType { PIE, DONUT; }
+
     Control createControl(Composite parent);
 
     void refresh(ClientSnapshot snapshot);
